likes

See merge request campuses/nlr/jan-2023/java-blue/student-pairs/module-3-final-capstone/nlr-12-module-3-final-capstone-party-playlist!98

+++ vue/src/components/EventDetail.vue
@@ -13,6 +13,7 @@
     <input v-model="event.date" :disabled="!isDJ && !isHost" />
     <label>Event Theme</label>
     <input v-model="event.theme" :disabled="!isDJ && !isHost" />
+    <button @click="submitSong(song.id, event.playlist.id)">Submit</button>
     <!-- <div v-if="isLoading">Loading...</div> -->
     <h1>{{ event.name }}</h1>
     <p>{{ event.description }}</p>
@@ -26,42 +27,20 @@
     <p>{{ event.djUsername }}</p>
     <p>Playlist: {{ event.playlist.name }}</p>
     <div class="song-info">
-      <song-display
-        v-for="song in event.playlist.songs"
-        :key="song.song_id"
-        :song="song"
-      >
+      <song-display v-for="song in event.playlist.songs" :key="song.song_id" :song="song">
         <div style="display: flex; flex-direction: column; justify-content: space-between">
           <div id="likes">
             <span>{{ song.likes }}</span>
-            <img
-              src="../images/thumbs-up.png"
-              alt="Likes"
-              width="15"
-              height="15"
-              class="thumb"
-              style="margin-bottom: 10px; cursor: pointer"
-              @click="incrementLikes(song.id)"
-              :class="{ disabled: song.clicked }"
-            />
+            <img src="../images/thumbs-up.png" alt="Likes" width="15" height="15" class="thumb" style="margin-bottom: 10px; cursor: pointer" @click="incrementLikes(song.id)" :class="{ disabled: song.clicked || clickedSongs.includes(song.id) }"/>
           </div>
           <div id="dislikes">
             <span>{{ song.dislikes }}</span>
-            <img
-            src="../images/thumbs-down.png"
-            alt="Dislikes"
-            width="15"
-            height="15"
-            class="thumb"
-            style="margin-bottom: 10px; cursor: pointer"
-            @click="decrementLikes(song.id)"
-            :class="{ disabled: song.clicked }"
-          />
+            <img src="../images/thumbs-down.png" alt="Dislikes" width="15" height="15" class="thumb" style="margin-bottom: 10px; cursor: pointer" @click="decrementLikes(song.id)" :class="{ disabled: song.clicked || clickedSongs.includes(song.id) }"/>
           </div>
         </div>
       </song-display>
     </div>
-    <button @click="submitSong(song.id, event.playlist.id)">Submit</button>
+
   </div>
 
 
@@ -81,13 +60,14 @@ export default {
   },
   data() {
     return {
-      user: null,
-      isVisible: false,
-      isLoading: true,
-      event: {},
-      error: "",
-    }
-  },
+    user: null,
+    isVisible: false,
+    isLoading: true,
+    event: {},
+    error: "",
+    clickedSongs: [], 
+  }
+},
   created() {
     this.getEvent();
   },
@@ -114,13 +94,21 @@ export default {
         });
     },
     incrementLikes(songId) {
-      playlistService.addLikes(this.event.playlist.playlistId, songId);
-      this.getEvent();
-    },
-    decrementLikes(songId) {
-      playlistService.deleteLikes(this.event.playlist.playlistId, songId);
-      this.getEvent();
-    },
+  if (this.clickedSongs.includes(songId)) {
+    return;
+  }
+  playlistService.addLikes(this.event.playlist.playlistId, songId);
+  this.clickedSongs.push(songId); // add this line
+  this.getEvent();
+},
+decrementLikes(songId) {
+  if (this.clickedSongs.includes(songId)) {
+    return;
+  }
+  playlistService.deleteLikes(this.event.playlist.playlistId, songId);
+  this.clickedSongs.push(songId); // add this line
+  this.getEvent();
+},
   },
 submitSong(songId, playlistId) {
     axios.post('/api/add-song-to-playlist', {
@@ -141,7 +129,10 @@ submitSong(songId, playlistId) {
 }
 .disabled {
   opacity: 0.2;
-  pointer-events: none;
+  cursor: not-allowed;
+}
+.disabled img {
+  filter: grayscale(1);
 }
 #likes {
   display: flex;
